Push: Added support for multifile editor, fix #93

+++ src/commands/push.js
@@ -1,9 +1,11 @@
 const fs = require('fs');
 const { ApifyCommand } = require('../lib/apify_command');
 const { flags: flagsHelper } = require('@oclif/command');
-const { createActZip, getLocalConfigOrThrow,
-    getLoggedClientOrThrow, outputJobLog, getLocalUserInfo } = require('../lib/utils');
-const { ACT_JOB_STATUSES, ACT_SOURCE_TYPES } = require('apify-shared/consts');
+const { createActZip, getLoggedClientOrThrow,
+    outputJobLog, getLocalUserInfo, getActorLocalFilePaths,
+    createSourceFiles, getLocalConfigOrThrow } = require('../lib/utils');
+const { sumFilesSizeInBytes } = require('../lib/files');
+const { ACT_JOB_STATUSES, ACT_SOURCE_TYPES, MAX_MULTIFILE_BYTES } = require('apify-shared/consts');
 const { DEFAULT_ACT_TEMPLATE, ACTS_TEMPLATES, UPLOADS_STORE_NAME } = require('../lib/consts');
 const { transformEnvToEnvVars } = require('../lib/secrets');
 const outputs = require('../lib/outputs');
@@ -57,27 +59,39 @@ class PushCommand extends ApifyCommand {
 
         outputs.info(`Deploying actor '${localConfig.name}' to Apify.`);
 
-        // Create zip
-        outputs.run('Zipping actor files');
-        await createActZip(TEMP_ZIP_FILE_NAME);
+        const filePathsToPush = await getActorLocalFilePaths();
+        const filesSize = await sumFilesSizeInBytes(filePathsToPush);
 
-        // Upload it to Apify.keyValueStores
-        const store = await apifyClient.keyValueStores.getOrCreateStore({ storeName: UPLOADS_STORE_NAME });
-        const key = `${actor.name}-${version}.zip`;
-        const buffer = fs.readFileSync(TEMP_ZIP_FILE_NAME);
-        await apifyClient.keyValueStores.putRecord({
-            storeId: store.id,
-            key,
-            body: buffer,
-            contentType: 'application/zip',
-        });
-        fs.unlinkSync(TEMP_ZIP_FILE_NAME);
-        const tarballUrl = `https://api.apify.com/v2/key-value-stores/${store.id}/records/${key}?disableRedirect=true`;
+        let sourceType;
+        let sourceFiles;
+        let tarballUrl;
+        if (filesSize < MAX_MULTIFILE_BYTES) {
+            sourceFiles = await createSourceFiles(filePathsToPush);
+            sourceType = ACT_SOURCE_TYPES.SOURCE_FILES;
+        } else {
+            // Create zip
+            outputs.run('Zipping actor files');
+            await createActZip(TEMP_ZIP_FILE_NAME, filePathsToPush);
+
+            // Upload it to Apify.keyValueStores
+            const store = await apifyClient.keyValueStores.getOrCreateStore({ storeName: UPLOADS_STORE_NAME });
+            const key = `${actor.name}-${version}.zip`;
+            const buffer = fs.readFileSync(TEMP_ZIP_FILE_NAME);
+            await apifyClient.keyValueStores.putRecord({
+                storeId: store.id,
+                key,
+                body: buffer,
+                contentType: 'application/zip',
+            });
+            fs.unlinkSync(TEMP_ZIP_FILE_NAME);
+            tarballUrl = `https://api.apify.com/v2/key-value-stores/${store.id}/records/${key}?disableRedirect=true`;
+            sourceType = ACT_SOURCE_TYPES.TARBALL;
+        }
 
         // Update actor version
         const actorCurrentVersion = await apifyClient.acts.getActVersion({ actId: actorId, versionNumber: version });
         if (actorCurrentVersion) {
-            const actorVersionModifier = { tarballUrl, buildTag, sourceType: ACT_SOURCE_TYPES.TARBALL };
+            const actorVersionModifier = { tarballUrl, sourceFiles, buildTag, sourceType };
             if (localConfig.env) actorVersionModifier.envVars = transformEnvToEnvVars(localConfig.env);
             await apifyClient.acts.updateActVersion({
                 actId: actorId,
@@ -89,8 +103,9 @@ class PushCommand extends ApifyCommand {
             const actorNewVersion = {
                 versionNumber: version,
                 tarballUrl,
+                sourceFiles,
                 buildTag,
-                sourceType: ACT_SOURCE_TYPES.TARBALL,
+                sourceType,
             };
             if (localConfig.env) actorNewVersion.envVars = transformEnvToEnvVars(localConfig.env);
             await apifyClient.acts.createActVersion({
@@ -142,7 +157,7 @@ PushCommand.flags = {
         description: 'DEPRECATED: Use flag version instead. Actor version number to which the files should be pushed. By default, it is taken from the "apify.json" file.',
         required: false,
     }),
-    'version': flagsHelper.string({
+    version: flagsHelper.string({
         char: 'v',
         description: 'Actor version number to which the files should be pushed. By default, it is taken from the "apify.json" file.',
         required: false,
@@ -163,8 +178,8 @@ PushCommand.args = [
     {
         name: 'actorId',
         required: false,
-        description: 'ID of an existing actor on the Apify platform where the files will be pushed. ' +
-        'If not provided, the command will create or modify the actor with the name specified in "apify.json" file.',
+        description: 'ID of an existing actor on the Apify platform where the files will be pushed. '
+        + 'If not provided, the command will create or modify the actor with the name specified in "apify.json" file.',
     },
 ];
